+++ test/test_brood.rb
@@ -19,11 +19,11 @@ describe Brood do
         store = @brood.instance_variable_get(:@store)
         refute_nil(store)
         assert_equal(2, store.dig(:department, :objects).length)
-        assert_equal(7, store.dig(:user, :objects).length)
+        assert_equal(8, store.dig(:user, :objects).length)
       end
     end
 
-    describe "#set" do
+    describe "#create" do
       it "creates an object" do
         department_count = -> {
           @brood.instance_variable_get(:@store).dig(:department, :objects).size
@@ -33,15 +33,15 @@ describe Brood do
         }
 
         assert_equal(2, department_count.call)
-        assert_equal(7, user_count.call)
+        assert_equal(8, user_count.call)
 
-        @brood.set(%i[user jimbo],
+        @brood.create(%i[user jimbo],
           {
             id: 1_234,
             name: "Jimbo"
           })
 
-        assert_equal(8, user_count.call)
+        assert_equal(9, user_count.call)
 
         user_objects = @brood.instance_variable_get(:@store).dig(:user, :objects)
         user_objects.each do |object_name, object|
@@ -56,25 +56,85 @@ describe Brood do
 
       it "raises an exception on invalid arguments" do
         assert_raises(ArgumentError) {
-          @brood.set([:user, :foo]) {}
+          @brood.create([:user, :foo]) {}
         }
         assert_raises(ArgumentError) {
-          @brood.set(["user", :baar], {id: 21_322})
+          @brood.create(["user", :baar], {id: 21_322})
         }
         assert_raises(ArgumentError) {
-          @brood.set([:user, "foo"], {id: 23_420})
+          @brood.create([:user, "foo"], {id: 23_420})
         }
         assert_raises(TypeError) {
-          @brood.set([:user, :baaz], 2432)
+          @brood.create([:user, :baaz], 2432)
         }
         assert_raises(TypeError) {
-          @brood.set([:user, :baaz], "xyz")
+          @brood.create([:user, :baaz], "xyz")
         }
       end
 
       it "raises an exception when an object is already defined for the identifier" do
         assert_raises(Brood::ObjectAlreadyDefinedError) do
-          @brood.set(%i[user alice],
+          @brood.create(%i[user alice],
+            {
+              id: 5_325,
+              name: "Alice"
+            })
+        end
+      end
+    end
+
+    describe "#build" do
+      it "builds an object" do
+        department_count = -> {
+          @brood.instance_variable_get(:@store).dig(:department, :objects).size
+        }
+        user_count = -> {
+          @brood.instance_variable_get(:@store).dig(:user, :objects).size
+        }
+
+        assert_equal(2, department_count.call)
+        assert_equal(8, user_count.call)
+
+        @brood.build(%i[user jimbo],
+          {
+            id: 1_234,
+            name: "Jimbo"
+          })
+
+        assert_equal(9, user_count.call)
+
+        user_objects = @brood.instance_variable_get(:@store).dig(:user, :objects)
+        user_objects.each do |object_name, object|
+          assert_instance_of(Symbol, object_name)
+          assert_instance_of(User, object)
+        end
+        assert_instance_of(Department, @brood.get(%i[department widgets_qc]))
+        assert_instance_of(Department, @brood.get(%i[department gizmos_qc]))
+        assert_instance_of(User, @brood.get(%i[user jimbo]))
+        assert_equal("Jimbo", @brood.get(%i[user jimbo]).name)
+      end
+
+      it "raises an exception on invalid arguments" do
+        assert_raises(ArgumentError) {
+          @brood.build([:user, :foo]) {}
+        }
+        assert_raises(ArgumentError) {
+          @brood.build(["user", :baar], {id: 21_322})
+        }
+        assert_raises(ArgumentError) {
+          @brood.build([:user, "foo"], {id: 23_420})
+        }
+        assert_raises(TypeError) {
+          @brood.build([:user, :baaz], 2432)
+        }
+        assert_raises(TypeError) {
+          @brood.build([:user, :baaz], "xyz")
+        }
+      end
+
+      it "raises an exception when an object is already defined for the identifier" do
+        assert_raises(Brood::ObjectAlreadyDefinedError) do
+          @brood.build(%i[user alice],
             {
               id: 5_325,
               name: "Alice"
@@ -84,8 +144,13 @@ describe Brood do
     end
 
     describe "#get" do
-      it "retrieves an object from the store" do
+      it "retrieves an object instantiated with 'create' from the store" do
         object = @brood.get(%i[user alice])
+        assert_instance_of(User, object)
+      end
+
+      it "retrieves an object instantiated with 'build' from the store" do
+        object = @brood.get(%i[user foobar])
         assert_instance_of(User, object)
       end