+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.2.0] - 2024-07-18
+### Replace method :set with :create and :build
+
+Features:
+
+* Add method `Brood#create` (calls `Fabricate`)
+* Add method `Brood#build` (calls `Fabricate.build`)
+
+Breaking changes:
+
+* Remove method `Brood#set`
+
+Misc:
+
+* Update README.md
+
 ## [0.1.0] - 2024-07-18
 ### Initial release